BREAK EVERYTHING ;)
created an itemview base class which can be used in quickaccess and
quickstarter. plasmoidspecific functions(contextmenu, dnd, etc.) should
go into a derived class(ItemView). I introduced a base class for
animations(Animator) which should make it possible to easily write
animations for the itemview. Atm there is a fadeanimation(the default
till now) and a flip animation known from kickoff, but I have an idea
for at least one other animation...
This is work in progress, so don't wonder if it is still buggy ;)